alphabet sort fix

+++ diplom/lib/Services/Data.dart
@@ -109,7 +109,7 @@ class Lesson {
     //if first in list
     } else if (id == courseLessons.first.id) {
       return LessonStateTypes.Current;
-    } else if (data.user.completedLessonsID.contains(courseLessons[max<int>(0,courseLessons.indexOf(courseLessons.where((element) => element.id == id).first)-1 as int)].id)) {
+    } else if (data.user.completedLessonsID.contains(courseLessons[max<int>(0,courseLessons.indexOf(courseLessons.where((element) => element.id == id).first)-1)].id)) {
       // if (data.user.completedLessonsID.contains(courseLessons.where((element) => element.id ==  (courseLessons.indexOf(courseLessons.where((element) => element.id == id).first)-1)).first.id)) {
       //   return 
       // }
